refactor: Move count -> sourcenode/count
