+++ internal/pkg/service/stream/storage/level/local/writer/sourcenode/count/counter.go
